Backend/requirements.txt
flask
flask-restful
flask-marshmallow
web3